requirements.txt
python-telegram-bot==20.7
stripe==8.5.0
flask==2.3.3